created modal

+++ client/todo/src/components/Todo.js
@@ -1,16 +1,24 @@
-import React, { useCallback, useMemo, useState } from 'react'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { TextField, useTheme } from '@mui/material'
 import { useDispatch } from 'react-redux'
 import { deleteTodoRequest, updateTodoRequest } from '../store/actions/todos.js'
 
-function Todo(props) {
+function Todo({
+  todo,
+  editingTodoId,
+  handleCurrentTodo,
+  setOpen,
+  isDelete,
+  setDeleteTodoId,
+  deleteTodoId,
+  setIsDelete,
+}) {
   const dispatch = useDispatch()
   const theme = useTheme()
   const deleteTodoAction = (payload) => dispatch(deleteTodoRequest(payload))
   const updateTodoAction = (payload) => dispatch(updateTodoRequest(payload))
   const [inputValue, setInputValue] = useState('')
   const [error, setError] = useState(false)
-  const { todo, editingTodoId, handleCurrentTodo } = props
 
   const handleKeyDown = useCallback(() => {}, [])
   const editTodo = useCallback(({ target }) => {
@@ -44,8 +52,16 @@ function Todo(props) {
   }, [todo.active])
 
   const handleDeleteTodo = useCallback(() => {
-    deleteTodoAction(todo._id)
-  }, [])
+    setDeleteTodoId(todo._id)
+    setOpen(true)
+  }, [todo._id])
+
+  useEffect(() => {
+    if (isDelete && deleteTodoId === todo._id) {
+      deleteTodoAction(todo._id)
+      setIsDelete(false)
+    }
+  }, [isDelete])
 
   const handeEditingMode = useCallback(() => {
     const name = inputValue.trim()
